revert: Merges useSetLocation and useRegionModal into useRegionManager
