+++ packages/core/src/Layout.tsx
@@ -1,12 +1,12 @@
 import { useMemo, type PropsWithChildren, type ReactElement } from 'react'
 
-import { useRegion } from './components/region/RegionModal/useRegion'
+import { useRegionModal } from './components/region/RegionModal/useRegionModal'
 import { usePageViewEvent } from './sdk/analytics/hooks/usePageViewEvent'
 
 function Layout({ children }: PropsWithChildren) {
   const props = useMemo(() => (children as ReactElement)?.props, [])
   usePageViewEvent(props)
-  useRegion()
+  useRegionModal()
 
   return <>{children}</>
 }
